main: report version from Cargo.toml
This is slightly less informative than the `git describe`-derived
version introduced in #306 (for it's got nothing to do with git-derived
data), but it allows us to keep a single source of truth for the
version: Cargo.toml.

In fact, the version reported by `git describe` on master right now is
not 0.8.118, but 0.8.116-168-g8f11fd7 (168 commits since tag master),
because we're missing pushed tags for the other versions, it seems.

Signed-off-by: Stephan Renatus <srenatus@chef.io>