Skip if_/3 inlining if Then_0 is unsafe or invalid
As a part of optimization goal expansion in library(reif) inlines Then_0
argument verbatim into predicate body – this avoids unnecessary call/N
invocations and dramatically increases performance, but not all goals are safe
to be inlined in such a way. Here we are skipping this optimization if !s or
invalid goal were detected to prevent undesired side-effects from leaking into
outer goals.