Properly encapsulate `raw_block` to enforce its invariants
This notably tightens assertions in RawBlock::get to require indices to be less than head.
A special case for Machine::get_clause_p, as it needs to read a dangling stack frame.

Switch `RawBlock.head` to be a Cell to reduce the number of invariants and be closer to
being thread-safe.