Revert "FIXED: CLP(B): Delay BDD restriction until after the instantiation."
This reverts commit e185b626bdf1a99ab8188ae39875f645c0de90f0.

This change is now no longer needed, and the underlying issue is
apparently somewhere else entirely. See the description at:

    https://github.com/mthom/scryer-prolog/issues/2732

Current master behaves differently from Scryer as it was at
099d9aaca6509496ccb28c522659b990f42037f0 (i.e., preceding
the commit that is now being reverted), even on the same file.

For an example, see:

    https://github.com/triska/scryer-prolog/commit/dd41176b97d8369c3bb7de1c9932a9c3d10a0466

Scryer now works as expected, and compatibly with SICStus. We still
need to find out what fixed the root cause of this issue.