Rewrite scan_slice_to_str to be safe and correct, add AsHeapView trait to deduplicate unsafe code
scan_slice_to_str was close to being safe already, but the code using it needed to be unsafe,
and the safety of scan_slice_to_str was hinging on the correctness of the caller unsafe block.

It also wrongly assumed that Heap.inner.ptr was aligned to Heap::heap_cell_alignment().

Instead, I have rewritten `scan_slice_to_str` (and renamed it to the more accurate name
`scan_pstr_at`), to fix the incorrect computation caused by this wrong assumption,
and to reduce the unsafe blocks to their minimum, whose safety I could then prove.

The invariants of Heap remain to be fully vetted, as some of the code in this module remain
~wildly unsafe~ :)