Define ?- as an infix operator (xfx, 1200) for binary quad syntax
To support the binary quad syntax (Goal ?- Answer), we need ?- to be
recognized as an infix operator in addition to its existing prefix form.

This mirrors how :- is defined as both prefix and infix, allowing it to
be used for both directives (:- directive) and clauses (Head :- Body).

Now ?- can be used for:
- Prefix form (traditional): ?- query.
- Infix form (binary quads): goal ?- answer.